+++ index.html
@@ -57,7 +57,9 @@
 
                   <!-- Изменение глубины эффекта, накладываемого на изображение -->
                   <fieldset class="img-upload__effect-level  effect-level">
-                    <input class="effect-level__value" type="number" step="any" name="effect-level" value="">
+                    <label>
+                      <input class="effect-level__value" type="number" step="any" name="effect-level" value="">
+                    </label>
                     <div class="effect-level__slider"></div>
                   </fieldset>
 
@@ -116,10 +118,14 @@
                 <!-- Добавление хэш-тегов и комментария к изображению -->
                 <fieldset class="img-upload__text text">
                   <div class="img-upload__field-wrapper">
-                    <input class="text__hashtags" name="hashtags" placeholder="#ХэшТег">
+                    <label>
+                      <input class="text__hashtags" name="hashtags" placeholder="#ХэшТег">
+                    </label>
                   </div>
                   <div class="img-upload__field-wrapper">
-                    <textarea class="text__description" name="description" placeholder="Ваш комментарий..."></textarea>
+                    <label>
+                      <textarea class="text__description" name="description" placeholder="Ваш комментарий..."></textarea>
+                    </label>
                   </div>
                 </fieldset>
 
@@ -172,7 +178,9 @@
           <!-- Форма для отправки комментария -->
           <div class="social__footer">
             <img class="social__picture" src="img/avatar-6.svg" alt="Аватар комментатора фотографии" width="35" height="35">
-            <input type="text" class="social__footer-text" placeholder="Ваш комментарий...">
+            <label>
+              <input type="text" class="social__footer-text" placeholder="Ваш комментарий...">
+            </label>
             <button type="button" class="social__footer-btn" name="button">Отправить</button>
           </div>
         </div>
@@ -232,7 +240,9 @@
     <div class="img-upload__message  img-upload__message--loading">Загружаем...</div>
   </template>
 
-  <script src="js/main.js"></script>
+  <script src="js/main.js" type="module"></script>
+  <script src="js/data.js" type="module"></script>
+  <script src="js/util.js"></script>
 
 </body>
 </html>
